requirements.txt
flask
flask-sqlalchemy
python-dotenv
langchain
openai
faiss-cpu
redis
rq
requests
beautifulsoup4